just load images in viewport
The problem was that e.g. when the width was at 1920px the first 6 elements were frawn on screen for 0.1 seconds or something that triggered the loading of the first 6 images but only 3 were in the viewport. This problem was present for all layouts => 2 images visible then 4 get loaded and 1 image visible then 2 get loaded.